refactor: #264 - 관리자 공통 쿼리 유틸 import 구조 개선
