+++ src/features/admin/feedbacks/utils/feedback-user-search.ts
@@ -1,7 +1,7 @@
 import { createAdminClient } from "@/lib/supabase/admin";
 
+import { escapePostgrestLikePattern } from "../../utils/query";
 import { AdminFeedbackListQuery } from "../types/feedback-list";
-import { escapePostgrestLikePattern } from "./feedback-query";
 
 const UUID_PATTERN =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
